Gather (and email) more debug info.
- onUpgrade now copies the DB at the *start* of processing if it is an
upgrade, and gives it a unique name based on the original version.

- The help page now has a 'Send Information' button which will copy
the DB, and create a mail message with the DB and any log files
attached. The string, debug_email, can be a ';'-separated list of
email address to receive emails.

- The Help page now has a 'Cleanup Files' button which will cleanup
any log files or DB backups.

+++ src/com/eleybourn/bookcatalogue/Logger.java
@@ -66,14 +66,14 @@ public class Logger {
 	}
 
 	/**
-	 * Clear the error log each time the app is started
+	 * Clear the error log each time the app is started; preserve previous if non-empty
 	 */
 	public static void clearLog() {
 		try {
 			try { 
 				File orig = new File(Utils.ERRORLOG_FILE);
 				File backup = new File(Utils.ERRORLOG_FILE + ".bak");
-				if (orig.exists())
+				if (orig.exists() && orig.length() > 0)
 					orig.renameTo(backup);
 			} catch (Exception e) {
 				// Ignore backup failure...
